# Specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OTUS Highload Architect",
    "version": "1.2.0"
  },
  "paths": {
...
    "/post/create": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [ "text" ],
                "properties": {
                  "text": {
                    "$ref": "#/components/schemas/PostText"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно создан пост",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostId"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
    "/post/update": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [ "id", "text" ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/PostId"
                  },
                  "text": {
                    "$ref": "#/components/schemas/PostText"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно изменен пост"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "400": {
        "description": "Невалидные данные ввода"
      },
      "401": {
        "description": "Неавторизованный доступ"
      },
      "5xx": {
        "description": "Ошибка сервера",
        "headers": {
          "Retry-After": {
            "description": "Время, через которое еще раз нужно сделать запрос",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Описание ошибки"
                },
                "request_id": {
                  "type": "string",
                  "description": "Идентификатор запроса. Предназначен для более быстрого поиска проблем."
                },
                "code": {
                  "type": "integer",
                  "description": "Код ошибки. Предназначен для классификации проблем и более быстрого решения проблем."
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "PostId": {
        "type": "string",
        "description": "Идентификатор поста",
        "example": "1d535fd6-7521-4cb1-aa6d-031be7123c4d"
      },
      "PostText": {
        "type": "string",
        "description": "Текст поста",
        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris ultrices eros in cursus turpis massa."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Авторизация по токену, который был получен в методе /login"
      }
    }
  }
}